feat: implement GetSubscription endpoint and enhance AuthMiddleware to include user email
